Migrate non-CanvasKit-specific tests to ui/
