+++ engine/src/flutter/lib/web_ui/lib/src/engine/color_filter.dart
@@ -15,7 +15,7 @@ enum ColorFilterType { mode, matrix, linearToSrgbGamma, srgbToLinearGamma }
 ///
 /// Instances of this class are used with [Paint.colorFilter] on [Paint]
 /// objects.
-class EngineColorFilter implements SceneImageFilter, LayerImageFilter, ui.ColorFilter {
+class EngineColorFilter implements LayerImageFilter, ui.ColorFilter {
   /// Creates a color filter that applies the blend mode given as the second
   /// argument. The source color is the one given as the first argument, and the
   /// destination color is the one from the layer being composited.
@@ -129,4 +129,19 @@ class EngineColorFilter implements SceneImageFilter, LayerImageFilter, ui.ColorF
 
   @override
   Matrix4? get transform => null;
+
+  @override
+  bool operator ==(Object other) {
+    if (other is! EngineColorFilter) {
+      return false;
+    }
+    return other.type == type &&
+        other.color == color &&
+        other.blendMode == blendMode &&
+        listEquals(other.matrix, matrix);
+  }
+
+  @override
+  int get hashCode =>
+      Object.hash(type, color, blendMode, Object.hashAll(matrix ?? const <double>[]));
 }
