+++ engine/src/flutter/lib/web_ui/test/canvaskit/image_test.dart
@@ -31,46 +31,6 @@ void testMain() {
     image.dispose();
   });
 
-  test('Image constructor invokes onCreate once', () async {
-    int onCreateInvokedCount = 0;
-    ui.Image? createdImage;
-    ui.Image.onCreate = (ui.Image image) {
-      onCreateInvokedCount++;
-      createdImage = image;
-    };
-
-    final ui.Image image1 = await _createImage();
-
-    expect(onCreateInvokedCount, 1);
-    expect(createdImage, image1);
-
-    final ui.Image image2 = await _createImage();
-
-    expect(onCreateInvokedCount, 2);
-    expect(createdImage, image2);
-  });
-
-  test('dispose() invokes onDispose once', () async {
-    int onDisposeInvokedCount = 0;
-    ui.Image? disposedImage;
-    ui.Image.onDispose = (ui.Image image) {
-      onDisposeInvokedCount++;
-      disposedImage = image;
-    };
-
-    final ui.Image image1 = await _createImage()
-      ..dispose();
-
-    expect(onDisposeInvokedCount, 1);
-    expect(disposedImage, image1);
-
-    final ui.Image image2 = await _createImage()
-      ..dispose();
-
-    expect(onDisposeInvokedCount, 2);
-    expect(disposedImage, image2);
-  });
-
   test('fetchImage fetches image in chunks', () async {
     final List<int> cumulativeBytesLoadedInvocations = <int>[];
     final List<int> expectedTotalBytesInvocations = <int>[];
@@ -101,83 +61,6 @@ void testMain() {
 
     // Check the contents of the returned data.
     expect(result, List<int>.generate(100000, (int i) => i & 0xFF));
-  });
-
-  test('scaledImageSize scales to a target width with no target height', () {
-    final BitmapSize? size = scaledImageSize(200, 100, 600, null);
-    expect(size?.width, 600);
-    expect(size?.height, 300);
-  });
-
-  test('instantiateImageCodecFromBuffer dispose buffer', () async {
-    final ui.Image image = await _createImage();
-    final ByteData? imageData = await image.toByteData(format: ui.ImageByteFormat.png);
-    final ui.ImmutableBuffer imageBuffer = await ui.ImmutableBuffer.fromUint8List(
-      imageData!.buffer.asUint8List(),
-    );
-
-    final ui.Codec codec = await ui.instantiateImageCodecFromBuffer(imageBuffer);
-    codec.dispose();
-    image.dispose();
-
-    expect(imageBuffer.debugDisposed, isTrue);
-  });
-
-  test('instantiateImageCodecWithSize dispose buffer', () async {
-    // getTargetSize is null, so the image is not scaled.
-    final ui.Image image = await _createImage();
-    final ByteData? imageData = await image.toByteData(format: ui.ImageByteFormat.png);
-    final ui.ImmutableBuffer nullTargetSizeImageBuffer = await ui.ImmutableBuffer.fromUint8List(
-      imageData!.buffer.asUint8List(),
-    );
-
-    final ui.Codec codec = await ui.instantiateImageCodecWithSize(nullTargetSizeImageBuffer);
-    codec.dispose();
-    image.dispose();
-
-    expect(nullTargetSizeImageBuffer.debugDisposed, isTrue);
-
-    // getTargetSize is not null, so the image is scaled.
-    final ui.Image scaledImage = await _createImage();
-    final ByteData? scaledImageData = await scaledImage.toByteData(format: ui.ImageByteFormat.png);
-    final ui.ImmutableBuffer scaledImageBuffer = await ui.ImmutableBuffer.fromUint8List(
-      scaledImageData!.buffer.asUint8List(),
-    );
-
-    final ui.Codec scaledCodec = await ui.instantiateImageCodecWithSize(
-      scaledImageBuffer,
-      getTargetSize: (w, h) => ui.TargetImageSize(width: w ~/ 2, height: h ~/ 2),
-    );
-    scaledCodec.dispose();
-    scaledImage.dispose();
-
-    expect(scaledImageBuffer.debugDisposed, isTrue);
-  });
-
-  test('instantiateImageCodecWithSize disposes temporary image', () async {
-    final Set<ui.Image> activeImages = <ui.Image>{};
-    ui.Image.onCreate = activeImages.add;
-    ui.Image.onDispose = activeImages.remove;
-
-    final ui.Image image = await _createImage();
-    final ByteData? imageData = await image.toByteData(format: ui.ImageByteFormat.png);
-    final ui.ImmutableBuffer imageBuffer = await ui.ImmutableBuffer.fromUint8List(
-      imageData!.buffer.asUint8List(),
-    );
-    image.dispose();
-
-    final ui.Codec codec = await ui.instantiateImageCodecWithSize(
-      imageBuffer,
-      getTargetSize: (w, h) => ui.TargetImageSize(width: w ~/ 2, height: h ~/ 2),
-    );
-    final ui.FrameInfo frameInfo = await codec.getNextFrame();
-
-    expect(activeImages.length, 1);
-
-    frameInfo.image.dispose();
-    codec.dispose();
-
-    expect(activeImages.length, 0);
   });
 
   test('CkImage does not close image source too early', () async {
