+++ engine/src/flutter/lib/web_ui/test/canvaskit/native_memory_test.dart
@@ -6,7 +6,6 @@ import 'dart:js_interop';
 
 import 'package:test/bootstrap/browser.dart';
 import 'package:test/test.dart';
-
 import 'package:ui/src/engine.dart';
 
 import 'common.dart';
