+++ engine/src/flutter/lib/web_ui/test/canvaskit/painting_test.dart
@@ -4,7 +4,6 @@
 
 import 'package:test/bootstrap/browser.dart';
 import 'package:test/test.dart';
-
 import 'package:ui/src/engine.dart';
 
 import 'common.dart';
