+++ engine/src/flutter/lib/web_ui/test/canvaskit/picture_test.dart
@@ -6,7 +6,6 @@ import 'dart:typed_data';
 
 import 'package:test/bootstrap/browser.dart';
 import 'package:test/test.dart';
-
 import 'package:ui/src/engine.dart';
 import 'package:ui/ui.dart' as ui;
