+++ engine/src/flutter/lib/web_ui/test/ui/backdrop_filter_golden_test.dart
@@ -2,13 +2,18 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'dart:async';
+
 import 'package:test/bootstrap/browser.dart';
 import 'package:test/test.dart';
 import 'package:ui/src/engine.dart';
 import 'package:ui/ui.dart' as ui;
 import 'package:ui/ui_web/src/ui_web.dart' as ui_web;
+import 'package:web_engine_tester/golden_tester.dart';
 
-import 'common.dart';
+import '../common/rendering.dart';
+import '../common/test_initialization.dart';
+import 'utils.dart';
 
 void main() {
   internalBootstrapBrowserTest(() => testMain);
@@ -16,23 +21,18 @@ void main() {
 
 const ui.Rect region = ui.Rect.fromLTRB(0, 0, 500, 500);
 
-void testMain() {
-  group('BackdropFilter', () {
-    setUpCanvasKitTest(withImplicitView: true);
+Future<void> testMain() async {
+  setUpUnitTests(withImplicitView: true, setUpTestViewDimensions: false);
 
+  group('BackdropFilter', () {
     setUp(() {
       EngineFlutterDisplay.instance.debugOverrideDevicePixelRatio(1);
     });
 
-    tearDown(() {
-      PlatformViewManager.instance.debugClear();
-      CanvasKitRenderer.instance.debugClear();
-    });
-
     test('blur renders to the edges', () async {
       // Make a checkerboard picture so we can see the blur.
-      final CkPictureRecorder recorder = CkPictureRecorder();
-      final CkCanvas canvas = recorder.beginRecording(region);
+      final ui.PictureRecorder recorder = ui.PictureRecorder();
+      final ui.Canvas canvas = ui.Canvas(recorder, region);
       canvas.drawColor(const ui.Color(0xffffffff), ui.BlendMode.srcOver);
       final double sideLength = region.width / 20;
       final int rows = (region.height / sideLength).ceil();
@@ -45,29 +45,27 @@ void testMain() {
             sideLength,
             sideLength,
           );
-          canvas.drawRect(rect, CkPaint()..color = const ui.Color(0xffff0000));
+          canvas.drawRect(rect, ui.Paint()..color = const ui.Color(0xffff0000));
         }
       }
-      final CkPicture checkerboard = recorder.endRecording();
+      final ui.Picture checkerboard = recorder.endRecording();
 
-      final LayerSceneBuilder builder = LayerSceneBuilder();
+      final ui.SceneBuilder builder = ui.SceneBuilder();
       builder.pushOffset(0, 0);
       builder.addPicture(ui.Offset.zero, checkerboard);
       builder.pushBackdropFilter(ui.ImageFilter.blur(sigmaX: 10, sigmaY: 10));
-      await matchSceneGolden(
-        'canvaskit_backdropfilter_blur_edges.png',
-        builder.build(),
-        region: region,
-      );
+      await renderScene(builder.build());
+
+      await matchGoldenFile('ui_backdropfilter_blur_edges.png', region: region);
     });
     test('ImageFilter with ColorFilter as child', () async {
-      final LayerSceneBuilder builder = LayerSceneBuilder();
+      final ui.SceneBuilder builder = ui.SceneBuilder();
       const ui.Rect region = ui.Rect.fromLTRB(0, 0, 500, 250);
 
       builder.pushOffset(0, 0);
 
-      final CkPictureRecorder recorder = CkPictureRecorder();
-      final CkCanvas canvas = recorder.beginRecording(region);
+      final ui.PictureRecorder recorder = ui.PictureRecorder();
+      final ui.Canvas canvas = ui.Canvas(recorder, region);
       final ui.ColorFilter colorFilter = ui.ColorFilter.mode(
         const ui.Color(0XFF00FF00).withOpacity(0.55),
         ui.BlendMode.darken,
@@ -78,15 +76,13 @@ void testMain() {
       canvas.drawCircle(
         const ui.Offset(75, 125),
         50,
-        CkPaint()..color = const ui.Color.fromARGB(255, 255, 0, 0),
+        ui.Paint()..color = const ui.Color.fromARGB(255, 255, 0, 0),
       );
-      final CkPicture redCircle1 = recorder.endRecording();
+      final ui.Picture redCircle1 = recorder.endRecording();
       builder.addPicture(ui.Offset.zero, redCircle1);
-      await matchSceneGolden(
-        'canvaskit_red_circle_green_backdrop_colorFilter.png',
-        builder.build(),
-        region: region,
-      );
+
+      await renderScene(builder.build());
+      await matchGoldenFile('ui_red_circle_green_backdrop_colorFilter.png', region: region);
     });
 
     test('works with an invisible platform view inside', () async {
@@ -98,8 +94,8 @@ void testMain() {
       await createPlatformView(0, 'test-platform-view');
 
       // Make a checkerboard picture so we can see the blur.
-      final CkPictureRecorder recorder = CkPictureRecorder();
-      final CkCanvas canvas = recorder.beginRecording(region);
+      final ui.PictureRecorder recorder = ui.PictureRecorder();
+      final ui.Canvas canvas = ui.Canvas(recorder, region);
       canvas.drawColor(const ui.Color(0xffffffff), ui.BlendMode.srcOver);
       final double sideLength = region.width / 20;
       final int rows = (region.height / sideLength).ceil();
@@ -112,21 +108,21 @@ void testMain() {
             sideLength,
             sideLength,
           );
-          canvas.drawRect(rect, CkPaint()..color = const ui.Color(0xffff0000));
+          canvas.drawRect(rect, ui.Paint()..color = const ui.Color(0xffff0000));
         }
       }
-      final CkPicture checkerboard = recorder.endRecording();
+      final ui.Picture checkerboard = recorder.endRecording();
 
-      final LayerSceneBuilder builder = LayerSceneBuilder();
+      final ui.SceneBuilder builder = ui.SceneBuilder();
       builder.pushOffset(0, 0);
       builder.addPicture(ui.Offset.zero, checkerboard);
       builder.pushBackdropFilter(ui.ImageFilter.blur(sigmaX: 10, sigmaY: 10));
 
       // Draw a green rectangle, then an invisible platform view, then a blue
       // rectangle. Both rectangles should not be blurred.
-      final CkPictureRecorder greenRectRecorder = CkPictureRecorder();
-      final CkCanvas greenRectCanvas = greenRectRecorder.beginRecording(region);
-      final CkPaint greenPaint = CkPaint()..color = const ui.Color(0xff00ff00);
+      final ui.PictureRecorder greenRectRecorder = ui.PictureRecorder();
+      final ui.Canvas greenRectCanvas = ui.Canvas(greenRectRecorder, region);
+      final ui.Paint greenPaint = ui.Paint()..color = const ui.Color(0xff00ff00);
       greenRectCanvas.drawRect(
         ui.Rect.fromCenter(
           center: ui.Offset(region.width / 3, region.height / 2),
@@ -135,11 +131,11 @@ void testMain() {
         ),
         greenPaint,
       );
-      final CkPicture greenRectPicture = greenRectRecorder.endRecording();
+      final ui.Picture greenRectPicture = greenRectRecorder.endRecording();
 
-      final CkPictureRecorder blueRectRecorder = CkPictureRecorder();
-      final CkCanvas blueRectCanvas = blueRectRecorder.beginRecording(region);
-      final CkPaint bluePaint = CkPaint()..color = const ui.Color(0xff0000ff);
+      final ui.PictureRecorder blueRectRecorder = ui.PictureRecorder();
+      final ui.Canvas blueRectCanvas = ui.Canvas(blueRectRecorder, region);
+      final ui.Paint bluePaint = ui.Paint()..color = const ui.Color(0xff0000ff);
       blueRectCanvas.drawRect(
         ui.Rect.fromCenter(
           center: ui.Offset(2 * region.width / 3, region.height / 2),
@@ -148,7 +144,7 @@ void testMain() {
         ),
         bluePaint,
       );
-      final CkPicture blueRectPicture = blueRectRecorder.endRecording();
+      final ui.Picture blueRectPicture = blueRectRecorder.endRecording();
 
       builder.addPicture(ui.Offset.zero, greenRectPicture);
       builder.addPlatformView(0, width: 10, height: 10);
@@ -157,11 +153,9 @@ void testMain() {
       // Pop the backdrop filter layer.
       builder.pop();
 
-      await matchSceneGolden(
-        'canvaskit_backdropfilter_with_platformview.png',
-        builder.build(),
-        region: region,
-      );
+      await renderScene(builder.build());
+
+      await matchGoldenFile('canvaskit_backdropfilter_with_platformview.png', region: region);
     });
     // TODO(hterkelsen): https://github.com/flutter/flutter/issues/71520
   }, skip: isSafari || isFirefox);
