+++ engine/src/flutter/lib/web_ui/test/ui/platform_view_test.dart
@@ -13,6 +13,7 @@ import 'package:ui/ui_web/src/ui_web.dart' as ui_web;
 import 'package:web_engine_tester/golden_tester.dart';
 
 import '../common/rendering.dart';
+import '../common/test_data.dart';
 import '../common/test_initialization.dart';
 import 'utils.dart';
 
@@ -57,7 +58,7 @@ Future<void> testMain() async {
   });
 
   test('picture + overlapping platformView', () async {
-    await _createPlatformView(1, platformViewType);
+    await createPlatformView(1, platformViewType);
 
     final ui.PictureRecorder recorder = ui.PictureRecorder();
     final ui.Canvas canvas = ui.Canvas(recorder);
@@ -80,7 +81,7 @@ Future<void> testMain() async {
   });
 
   test('platformView sandwich', () async {
-    await _createPlatformView(1, platformViewType);
+    await createPlatformView(1, platformViewType);
 
     final ui.PictureRecorder recorder = ui.PictureRecorder();
     final ui.Canvas canvas = ui.Canvas(recorder);
@@ -107,7 +108,7 @@ Future<void> testMain() async {
   });
 
   test('transformed platformview', () async {
-    await _createPlatformView(1, platformViewType);
+    await createPlatformView(1, platformViewType);
 
     final ui.PictureRecorder recorder = ui.PictureRecorder();
     final ui.Canvas canvas = ui.Canvas(recorder);
@@ -131,7 +132,7 @@ Future<void> testMain() async {
   });
 
   test('transformed and offset platformview', () async {
-    await _createPlatformView(1, platformViewType);
+    await createPlatformView(1, platformViewType);
 
     final ui.PictureRecorder recorder = ui.PictureRecorder();
     final ui.Canvas canvas = ui.Canvas(recorder);
@@ -159,7 +160,7 @@ Future<void> testMain() async {
   });
 
   test('offset platformview', () async {
-    await _createPlatformView(1, platformViewType);
+    await createPlatformView(1, platformViewType);
 
     final ui.PictureRecorder recorder = ui.PictureRecorder();
     final ui.Canvas canvas = ui.Canvas(recorder);
@@ -194,7 +195,7 @@ Future<void> testMain() async {
   });
 
   test('platformview with opacity', () async {
-    await _createPlatformView(1, platformViewType);
+    await createPlatformView(1, platformViewType);
 
     final ui.PictureRecorder recorder = ui.PictureRecorder();
     final ui.Canvas canvas = ui.Canvas(recorder);
@@ -218,7 +219,7 @@ Future<void> testMain() async {
   });
 
   test('platformview cliprect', () async {
-    await _createPlatformView(1, platformViewType);
+    await createPlatformView(1, platformViewType);
 
     final ui.PictureRecorder recorder = ui.PictureRecorder();
     final ui.Canvas canvas = ui.Canvas(recorder);
@@ -242,7 +243,7 @@ Future<void> testMain() async {
   });
 
   test('platformview cliprrect', () async {
-    await _createPlatformView(1, platformViewType);
+    await createPlatformView(1, platformViewType);
 
     final ui.PictureRecorder recorder = ui.PictureRecorder();
     final ui.Canvas canvas = ui.Canvas(recorder);
@@ -269,7 +270,7 @@ Future<void> testMain() async {
   });
 
   test('platformview covered clip', () async {
-    await _createPlatformView(1, platformViewType);
+    await createPlatformView(1, platformViewType);
 
     final ui.PictureRecorder recorder = ui.PictureRecorder();
     final ui.Canvas canvas = ui.Canvas(recorder);
@@ -299,7 +300,7 @@ Future<void> testMain() async {
   });
 
   test('platformview clippath', () async {
-    await _createPlatformView(1, platformViewType);
+    await createPlatformView(1, platformViewType);
 
     final ui.PictureRecorder recorder = ui.PictureRecorder();
     final ui.Canvas canvas = ui.Canvas(recorder);
@@ -330,7 +331,7 @@ Future<void> testMain() async {
   });
 
   test('embeds interactive platform views', () async {
-    await _createPlatformView(1, platformViewType);
+    await createPlatformView(1, platformViewType);
 
     final ui.SceneBuilder sb = ui.SceneBuilder();
     sb.pushOffset(0, 0);
@@ -359,7 +360,7 @@ Future<void> testMain() async {
   });
 
   test('clips platform views with RRects', () async {
-    await _createPlatformView(1, platformViewType);
+    await createPlatformView(1, platformViewType);
 
     final ui.SceneBuilder sb = ui.SceneBuilder();
     sb.pushOffset(0, 0);
@@ -376,12 +377,32 @@ Future<void> testMain() async {
     );
     expect(sceneHost.querySelectorAll('flt-clip').single.style.width, '100%');
     expect(sceneHost.querySelectorAll('flt-clip').single.style.height, '100%');
-    // CanvasKit doesn't use the `rect` CSS clip-path to clip but it should:
-    // https://github.com/flutter/flutter/issues/172308
-  }, skip: isCanvasKit);
+  });
+
+  test('clips platform views with Paths', () async {
+    await createPlatformView(1, platformViewType);
+
+    final ui.SceneBuilder sb = ui.SceneBuilder();
+    sb.pushOffset(0, 0);
+    sb.pushClipPath(
+      ui.Path()
+        ..lineTo(0, 10)
+        ..lineTo(10, 0)
+        ..close(),
+    );
+    sb.addPlatformView(1, width: 10, height: 10);
+    await renderScene(sb.build());
+
+    expect(
+      sceneHost.querySelectorAll('flt-clip').single.style.clipPath,
+      'path("M 0 0 L 0 10 L 10 0 L 0 0 Z")',
+    );
+    expect(sceneHost.querySelectorAll('flt-clip').single.style.width, '100%');
+    expect(sceneHost.querySelectorAll('flt-clip').single.style.height, '100%');
+  });
 
   test('correctly transforms platform views', () async {
-    await _createPlatformView(1, platformViewType);
+    await createPlatformView(1, platformViewType);
 
     final ui.SceneBuilder sb = ui.SceneBuilder();
     sb.pushOffset(0, 0);
@@ -406,7 +427,7 @@ Future<void> testMain() async {
   });
 
   test('correctly offsets platform views', () async {
-    await _createPlatformView(1, platformViewType);
+    await createPlatformView(1, platformViewType);
 
     final ui.SceneBuilder sb = ui.SceneBuilder();
     sb.addPlatformView(1, offset: const ui.Offset(3, 4), width: 5, height: 6);
@@ -427,7 +448,7 @@ Future<void> testMain() async {
   });
 
   test('correctly offsets when clip chain length is changed', () async {
-    await _createPlatformView(1, platformViewType);
+    await createPlatformView(1, platformViewType);
 
     ui.SceneBuilder sb = ui.SceneBuilder();
     sb.pushOffset(3, 3);
@@ -462,14 +483,13 @@ Future<void> testMain() async {
       'matrix(1, 0, 0, 1, 6, 6)',
       'matrix(1, 0, 0, 1, 3, 3)',
     ]);
-    // Skwasm pre-transforms the clips instead of applying the transforms
-    // stepwise in each clip:
-    // https://github.com/flutter/flutter/issues/172308
+    // Unskip when Skwasm and CanvasKit are unified:
+    // https://github.com/flutter/flutter/issues/172311
   }, skip: isSkwasm);
 
   test('converts device pixels to logical pixels (no clips)', () async {
     EngineFlutterDisplay.instance.debugOverrideDevicePixelRatio(4);
-    await _createPlatformView(1, platformViewType);
+    await createPlatformView(1, platformViewType);
 
     final ui.SceneBuilder sb = ui.SceneBuilder();
     sb.pushOffset(1, 1);
@@ -483,13 +503,13 @@ Future<void> testMain() async {
 
     expect(getTransformChain(slotHost), <String>['matrix(0.25, 0, 0, 0.25, 1.5, 1.5)']);
     EngineFlutterDisplay.instance.debugOverrideDevicePixelRatio(null);
-    // Skwasm has an empty <flt-clip> element as a parent of the platform view:
-    // https://github.com/flutter/flutter/issues/172308
+    // Unskip when Skwasm and CanvasKit are unified:
+    // https://github.com/flutter/flutter/issues/172311
   }, skip: isSkwasm);
 
   test('converts device pixels to logical pixels (with clips)', () async {
     EngineFlutterDisplay.instance.debugOverrideDevicePixelRatio(4);
-    await _createPlatformView(1, platformViewType);
+    await createPlatformView(1, platformViewType);
 
     final ui.SceneBuilder sb = ui.SceneBuilder();
     sb.pushOffset(3, 3);
@@ -509,9 +529,8 @@ Future<void> testMain() async {
       'matrix(0.25, 0, 0, 0.25, 0.75, 0.75)',
     ]);
     EngineFlutterDisplay.instance.debugOverrideDevicePixelRatio(null);
-    // Skwasm pre-transforms the clips instead of applying the transforms
-    // stepwise in each clip:
-    // https://github.com/flutter/flutter/issues/172308
+    // Unskip when Skwasm and CanvasKit are unified:
+    // https://github.com/flutter/flutter/issues/172311
   }, skip: isSkwasm);
 
   test('renders overlays on top of platform views', () async {
@@ -526,7 +545,7 @@ Future<void> testMain() async {
     // Initialize all platform views to be used in the test.
     final List<int> platformViewIds = <int>[];
     for (int i = 0; i < 16; i++) {
-      await _createPlatformView(i, platformViewType);
+      await createPlatformView(i, platformViewType);
       platformViewIds.add(i);
     }
 
@@ -654,12 +673,12 @@ Future<void> testMain() async {
     // Frame 7:
     //   Render: a platform view after error.
     //   Expect: success. Just checking the system is not left in a corrupted state.
-    await _createPlatformView(0, platformViewType);
+    await createPlatformView(0, platformViewType);
     await renderTestScene(viewCount: 0);
     _expectSceneMatches(<_EmbeddedViewMarker>[]);
     debugOverrideJsConfiguration(null);
-    // Skwasm doesn't cap the maximum number of overlays:
-    // https://github.com/flutter/flutter/issues/172308
+    // Unskip when Skwasm and CanvasKit are unified:
+    // https://github.com/flutter/flutter/issues/172311
   }, skip: isSkwasm);
 
   test('correctly reuses overlays', () async {
@@ -671,7 +690,7 @@ Future<void> testMain() async {
     // Initialize all platform views to be used in the test.
     final List<int> platformViewIds = <int>[];
     for (int i = 0; i < 20; i++) {
-      await _createPlatformView(i, platformViewType);
+      await createPlatformView(i, platformViewType);
       platformViewIds.add(i);
     }
 
@@ -784,12 +803,12 @@ Future<void> testMain() async {
       _overlay,
       _platformView,
     ]);
-    // Skwasm doesn't cap the maximum number of overlays:
-    // https://github.com/flutter/flutter/issues/172308
+    // Unskip when Skwasm and CanvasKit are unified:
+    // https://github.com/flutter/flutter/issues/172311
   }, skip: isSkwasm);
 
   test('embeds and disposes of a platform view', () async {
-    await _createPlatformView(1, platformViewType);
+    await createPlatformView(1, platformViewType);
 
     ui.SceneBuilder sb = ui.SceneBuilder();
     sb.pushOffset(0, 0);
@@ -799,7 +818,7 @@ Future<void> testMain() async {
 
     expect(platformViewsHost.querySelector('flt-platform-view'), isNotNull);
 
-    await _disposePlatformView(1);
+    await disposePlatformView(1);
 
     sb = ui.SceneBuilder();
     sb.pushOffset(0, 0);
@@ -811,7 +830,7 @@ Future<void> testMain() async {
   });
 
   test('preserves the DOM node of an unrendered platform view', () async {
-    await _createPlatformView(1, platformViewType);
+    await createPlatformView(1, platformViewType);
 
     ui.SceneBuilder sb = ui.SceneBuilder();
     sb.pushOffset(0, 0);
@@ -822,7 +841,7 @@ Future<void> testMain() async {
     expect(platformViewsHost.querySelectorAll('flt-platform-view'), hasLength(1));
 
     // Render a frame with a different platform view.
-    await _createPlatformView(2, platformViewType);
+    await createPlatformView(2, platformViewType);
     sb = ui.SceneBuilder();
     sb.pushOffset(0, 0);
     sb.addPlatformView(2, width: 10, height: 10);
@@ -843,8 +862,51 @@ Future<void> testMain() async {
     expect(platformViewsHost.querySelectorAll('flt-platform-view'), hasLength(2));
   });
 
+  test(
+    'does not crash when resizing the window after textures have been registered',
+    () async {
+      ui_web.platformViewRegistry.registerViewFactory(
+        'test-platform-view',
+        (int viewId) => createDomHTMLDivElement()..id = 'view-0',
+      );
+      await createPlatformView(0, 'test-platform-view');
+
+      final ui.Codec codec = await renderer.instantiateImageCodec(kAnimatedGif);
+
+      final ui.FrameInfo frame = await codec.getNextFrame();
+      final ui.Image ckImage = frame.image;
+
+      final LayerSceneBuilder sb = LayerSceneBuilder();
+      sb.pushOffset(0, 0);
+      final ui.PictureRecorder recorder = ui.PictureRecorder();
+      final ui.Canvas canvas = ui.Canvas(recorder, ui.Rect.largest);
+      canvas.drawImage(ckImage, ui.Offset.zero, ui.Paint());
+      final ui.Picture picture = recorder.endRecording();
+      sb.addPicture(ui.Offset.zero, picture);
+      sb.addPlatformView(0, width: 10, height: 10);
+
+      implicitView.debugPhysicalSizeOverride = const ui.Size(100, 100);
+      implicitView.debugForceResize();
+      await renderScene(sb.build());
+      _expectSceneMatches(<_EmbeddedViewMarker>[_overlay, _platformView]);
+
+      implicitView.debugPhysicalSizeOverride = const ui.Size(200, 200);
+      implicitView.debugForceResize();
+      await renderScene(sb.build());
+      _expectSceneMatches(<_EmbeddedViewMarker>[_overlay, _platformView]);
+
+      implicitView.debugPhysicalSizeOverride = null;
+      implicitView.debugForceResize();
+
+      // ImageDecoder is not supported in Safari or Firefox.
+      // Unskip when Skwasm and CanvasKit are unified:
+      // https://github.com/flutter/flutter/issues/172311
+    },
+    skip: isSafari || isFirefox || isSkwasm,
+  );
+
   test('does not crash when a prerolled platform view is not composited', () async {
-    await _createPlatformView(1, platformViewType);
+    await createPlatformView(1, platformViewType);
 
     final ui.SceneBuilder sb = ui.SceneBuilder();
     sb.pushOffset(0, 0);
@@ -854,6 +916,8 @@ Future<void> testMain() async {
     // The below line should not throw an error.
     await renderScene(sb.build());
     _expectSceneMatches(<_EmbeddedViewMarker>[]);
+    // Unskip when Skwasm and CanvasKit are unified:
+    // https://github.com/flutter/flutter/issues/172311
   }, skip: isSkwasm);
 
   test('does not create overlays for invisible platform views', () async {
@@ -861,13 +925,13 @@ Future<void> testMain() async {
     final ui.Canvas testCanvas = ui.Canvas(testRecorder);
     testCanvas.drawCircle(const ui.Offset(5, 5), 5, ui.Paint());
     final ui.Picture testPicture = testRecorder.endRecording();
-    await _createPlatformView(0, platformViewType);
-    await _createPlatformView(1, invisiblePlatformViewType);
-    await _createPlatformView(2, platformViewType);
-    await _createPlatformView(3, invisiblePlatformViewType);
-    await _createPlatformView(4, invisiblePlatformViewType);
-    await _createPlatformView(5, invisiblePlatformViewType);
-    await _createPlatformView(6, invisiblePlatformViewType);
+    await createPlatformView(0, platformViewType);
+    await createPlatformView(1, invisiblePlatformViewType);
+    await createPlatformView(2, platformViewType);
+    await createPlatformView(3, invisiblePlatformViewType);
+    await createPlatformView(4, invisiblePlatformViewType);
+    await createPlatformView(5, invisiblePlatformViewType);
+    await createPlatformView(6, invisiblePlatformViewType);
 
     expect(PlatformViewManager.instance.isInvisible(0), isFalse);
     expect(PlatformViewManager.instance.isInvisible(1), isTrue);
@@ -1086,6 +1150,39 @@ Future<void> testMain() async {
       _platformView,
       _overlay,
     ]);
+    // Unskip when Skwasm and CanvasKit are unified:
+    // https://github.com/flutter/flutter/issues/172311
+  }, skip: isSkwasm);
+
+  test('can dispose without crashing', () async {
+    ui_web.platformViewRegistry.registerViewFactory(
+      'test-view',
+      (int viewId) => createDomHTMLDivElement()..className = 'platform-view',
+      isVisible: false,
+    );
+
+    await createPlatformView(0, 'test-view');
+    await createPlatformView(1, 'test-view');
+    await createPlatformView(2, 'test-view');
+
+    final ui.SceneBuilder sb = ui.SceneBuilder()
+      ..pushOffset(0, 0)
+      ..addPlatformView(0, width: 10, height: 10)
+      ..addPlatformView(1, width: 10, height: 10)
+      ..addPlatformView(2, width: 10, height: 10)
+      ..pop();
+
+    await renderScene(sb.build());
+
+    _expectSceneMatches(<_EmbeddedViewMarker>[_platformView, _platformView, _platformView]);
+
+    expect(() {
+      final PlatformViewEmbedder embedder = renderer.rasterizers[implicitView.viewId]!.viewEmbedder;
+      // The following line used to cause a "Concurrent modification during iteration"
+      embedder.dispose();
+    }, returnsNormally);
+    // Unskip when Skwasm and CanvasKit are unified:
+    // https://github.com/flutter/flutter/issues/172311
   }, skip: isSkwasm);
 
   test('optimizes overlays when pictures and platform views do not overlap', () async {
@@ -1096,9 +1193,9 @@ Future<void> testMain() async {
       return testRecorder.endRecording();
     }
 
-    await _createPlatformView(0, platformViewType);
-    await _createPlatformView(1, platformViewType);
-    await _createPlatformView(2, platformViewType);
+    await createPlatformView(0, platformViewType);
+    await createPlatformView(1, platformViewType);
+    await createPlatformView(2, platformViewType);
 
     expect(PlatformViewManager.instance.isVisible(0), isTrue);
     expect(PlatformViewManager.instance.isVisible(1), isTrue);
@@ -1227,8 +1324,8 @@ Future<void> testMain() async {
       return testRecorder.endRecording();
     }
 
-    await _createPlatformView(0, platformViewType);
-    await _createPlatformView(1, platformViewType);
+    await createPlatformView(0, platformViewType);
+    await createPlatformView(1, platformViewType);
 
     expect(PlatformViewManager.instance.isVisible(0), isTrue);
     expect(PlatformViewManager.instance.isVisible(1), isTrue);
@@ -1266,7 +1363,7 @@ Future<void> testMain() async {
       return testRecorder.endRecording();
     }
 
-    await _createPlatformView(0, platformViewType);
+    await createPlatformView(0, platformViewType);
 
     expect(PlatformViewManager.instance.isVisible(0), isTrue);
 
@@ -1282,30 +1379,266 @@ Future<void> testMain() async {
     await renderScene(scene);
     _expectSceneMatches(<_EmbeddedViewMarker>[_overlay, _platformView]);
   });
-}
 
-// Sends a platform message to create a Platform View with the given id and viewType.
-Future<void> _createPlatformView(int id, String viewType) {
-  final Completer<void> completer = Completer<void>();
-  const MethodCodec codec = StandardMethodCodec();
-  ui.PlatformDispatcher.instance.sendPlatformMessage(
-    'flutter/platform_views',
-    codec.encodeMethodCall(MethodCall('create', <String, dynamic>{'id': id, 'viewType': viewType})),
-    (dynamic _) => completer.complete(),
-  );
-  return completer.future;
-}
+  test('correctly places pictures in case where next '
+      'picture intersects multiple elements', () async {
+    ui_web.platformViewRegistry.registerViewFactory(
+      'test-view',
+      (int viewId) => createDomHTMLDivElement()..className = 'platform-view',
+    );
+    ui_web.platformViewRegistry.registerViewFactory(
+      'invisible-view',
+      (int viewId) => createDomHTMLDivElement()..className = 'invisible-platform-view',
+      isVisible: false,
+    );
 
-/// Disposes of the platform view with the given [id].
-Future<void> _disposePlatformView(int id) {
-  final Completer<void> completer = Completer<void>();
-  const MethodCodec codec = StandardMethodCodec();
-  ui.PlatformDispatcher.instance.sendPlatformMessage(
-    'flutter/platform_views',
-    codec.encodeMethodCall(MethodCall('dispose', id)),
-    (dynamic _) => completer.complete(),
-  );
-  return completer.future;
+    ui.Picture rectPicture(ui.Rect rect) {
+      final ui.PictureRecorder testRecorder = ui.PictureRecorder();
+      final ui.Canvas canvas = ui.Canvas(testRecorder);
+      canvas.drawRect(rect, ui.Paint()..color = const ui.Color.fromARGB(255, 255, 0, 0));
+      return testRecorder.endRecording();
+    }
+
+    await createPlatformView(0, 'test-view');
+    await createPlatformView(1, 'invisible-view');
+
+    expect(PlatformViewManager.instance.isVisible(0), isTrue);
+    expect(PlatformViewManager.instance.isVisible(1), isFalse);
+
+    final LayerSceneBuilder sb = LayerSceneBuilder();
+    sb.pushOffset(0, 0);
+    sb.addPicture(ui.Offset.zero, rectPicture(const ui.Rect.fromLTWH(0, 0, 100, 100)));
+    sb.addPlatformView(0, offset: const ui.Offset(10, 10), width: 50, height: 50);
+    sb.addPicture(ui.Offset.zero, rectPicture(const ui.Rect.fromLTWH(0, 0, 100, 100)));
+    sb.addPlatformView(1, offset: const ui.Offset(10, 10), width: 50, height: 50);
+    sb.addPicture(ui.Offset.zero, rectPicture(const ui.Rect.fromLTWH(0, 0, 5, 5)));
+    final LayerScene scene = sb.build();
+    await renderScene(scene);
+    _expectSceneMatches(<_EmbeddedViewMarker>[_overlay, _platformView, _platformView, _overlay]);
+
+    final Composition composition =
+        renderer.rasterizers[implicitView.viewId]!.viewEmbedder.debugActiveComposition;
+    final List<int> picturesPerCanvas = composition.canvases
+        .map((CompositionCanvas canvas) => canvas.pictures.length)
+        .toList();
+    expect(picturesPerCanvas, <int>[1, 2]);
+    // Unskip when Skwasm and CanvasKit are unified:
+    // https://github.com/flutter/flutter/issues/172311
+  }, skip: isSkwasm);
+
+  test('can customize amount of overlays', () async {
+    final ui.PictureRecorder testRecorder = ui.PictureRecorder();
+    final ui.Canvas testCanvas = ui.Canvas(testRecorder);
+    testCanvas.drawCircle(const ui.Offset(5, 5), 5, ui.Paint());
+    final ui.Picture testPicture = testRecorder.endRecording();
+
+    // Initialize all platform views to be used in the test.
+    final List<int> platformViewIds = <int>[];
+    for (int i = 0; i < 16; i++) {
+      ui_web.platformViewRegistry.registerViewFactory(
+        'test-platform-view',
+        (int viewId) => createDomHTMLDivElement()..id = 'view-$i',
+      );
+      await createPlatformView(i, 'test-platform-view');
+      platformViewIds.add(i);
+    }
+
+    Future<void> renderTestScene({required int viewCount}) async {
+      final LayerSceneBuilder sb = LayerSceneBuilder();
+      sb.pushOffset(0, 0);
+      for (int i = 0; i < viewCount; i++) {
+        sb.addPicture(ui.Offset.zero, testPicture);
+        sb.addPlatformView(i, width: 10, height: 10);
+      }
+      await renderScene(sb.build());
+    }
+
+    // Set maximum overlays to 4.
+    debugOverrideJsConfiguration(
+      <String, Object?>{'canvasKitMaximumSurfaces': 4}.jsify() as JsFlutterConfiguration?,
+    );
+
+    await renderTestScene(viewCount: 8);
+    _expectSceneMatches(<_EmbeddedViewMarker>[
+      _overlay,
+      _platformView,
+      _overlay,
+      _platformView,
+      _overlay,
+      _platformView,
+      _platformView,
+      _platformView,
+      _platformView,
+      _platformView,
+      _overlay,
+      _platformView,
+    ]);
+
+    // Set maximum overlays to -1. Should default to 1.
+    debugOverrideJsConfiguration(
+      <String, Object?>{'canvasKitMaximumSurfaces': -1}.jsify() as JsFlutterConfiguration?,
+    );
+
+    await renderTestScene(viewCount: 8);
+    _expectSceneMatches(<_EmbeddedViewMarker>[
+      _platformView,
+      _platformView,
+      _platformView,
+      _platformView,
+      _platformView,
+      _platformView,
+      _platformView,
+      _overlay,
+      _platformView,
+    ]);
+
+    debugOverrideJsConfiguration(null);
+    // Unskip when Skwasm and CanvasKit are unified:
+    // https://github.com/flutter/flutter/issues/172311
+  }, skip: isSkwasm);
+
+  test('correctly rearranges pictures to second-to-last canvas '
+      'when hitting canvas limit', () async {
+    final ui.PictureRecorder testRecorder = ui.PictureRecorder();
+    final ui.Canvas testCanvas = ui.Canvas(testRecorder);
+    testCanvas.drawCircle(const ui.Offset(5, 5), 5, ui.Paint());
+    final ui.Picture testPicture = testRecorder.endRecording();
+
+    // Initialize all platform views to be used in the test.
+    final List<int> platformViewIds = <int>[];
+    for (int i = 0; i < 20; i++) {
+      ui_web.platformViewRegistry.registerViewFactory(
+        'test-platform-view',
+        (int viewId) => createDomHTMLDivElement()..id = 'view-$i',
+      );
+      await createPlatformView(i, 'test-platform-view');
+      platformViewIds.add(i);
+    }
+
+    Future<void> renderTestScene(List<int> views) async {
+      final ui.SceneBuilder sb = ui.SceneBuilder();
+      sb.pushOffset(0, 0);
+      for (final int view in views) {
+        sb.addPicture(ui.Offset.zero, testPicture);
+        sb.addPlatformView(view, width: 10, height: 10);
+      }
+      await renderScene(sb.build());
+    }
+
+    // Render scene with 20 pictures. Check that the second-to-last canvas
+    // contains the pictures from the canvases that were deleted.
+    await renderTestScene(<int>[1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19]);
+    _expectSceneMatches(<_EmbeddedViewMarker>[
+      _overlay,
+      _platformView,
+      _overlay,
+      _platformView,
+      _overlay,
+      _platformView,
+      _overlay,
+      _platformView,
+      _overlay,
+      _platformView,
+      _overlay,
+      _platformView,
+      _overlay,
+      _platformView,
+      _platformView,
+      _platformView,
+      _platformView,
+      _platformView,
+      _platformView,
+      _platformView,
+      _platformView,
+      _platformView,
+      _platformView,
+      _platformView,
+      _platformView,
+      _overlay,
+      _platformView,
+    ]);
+
+    // The second-to-last canvas should have all the extra pictures.
+    final Composition composition =
+        renderer.rasterizers[implicitView.viewId]!.viewEmbedder.debugActiveComposition;
+    final List<int> numPicturesPerCanvas = composition.canvases
+        .map((CompositionCanvas canvas) => canvas.pictures.length)
+        .toList();
+    expect(numPicturesPerCanvas, <int>[1, 1, 1, 1, 1, 1, 12, 1]);
+
+    // It should also work when the maximum canvases is just one.
+    debugOverrideJsConfiguration(
+      <String, Object?>{'canvasKitMaximumSurfaces': 1}.jsify() as JsFlutterConfiguration?,
+    );
+
+    // Render scene with 20 pictures. Check that the last canvas contains the
+    // pictures from the canvases that were deleted.
+    await renderTestScene(<int>[1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19]);
+    _expectSceneMatches(<_EmbeddedViewMarker>[
+      _platformView,
+      _platformView,
+      _platformView,
+      _platformView,
+      _platformView,
+      _platformView,
+      _platformView,
+      _platformView,
+      _platformView,
+      _platformView,
+      _platformView,
+      _platformView,
+      _platformView,
+      _platformView,
+      _platformView,
+      _platformView,
+      _platformView,
+      _platformView,
+      _overlay,
+      _platformView,
+    ]);
+
+    // The last canvas should have all the pictures.
+    final Composition secondComposition =
+        renderer.rasterizers[implicitView.viewId]!.viewEmbedder.debugActiveComposition;
+    final List<int> picturesPerCanvasInSecondRendering = secondComposition.canvases
+        .map((CompositionCanvas canvas) => canvas.pictures.length)
+        .toList();
+    expect(picturesPerCanvasInSecondRendering, <int>[19]);
+    debugOverrideJsConfiguration(null);
+    // Unskip when Skwasm and CanvasKit are unified:
+    // https://github.com/flutter/flutter/issues/172311
+  }, skip: isSkwasm);
+
+  test('disposes render pictures', () async {
+    Instrumentation.enabled = true;
+    Instrumentation.instance.debugCounters.clear();
+
+    ui_web.platformViewRegistry.registerViewFactory(
+      'test-view',
+      (int viewId) => createDomHTMLDivElement()..className = 'platform-view',
+    );
+
+    ui.Picture rectPicture(ui.Rect rect) {
+      final ui.PictureRecorder testRecorder = ui.PictureRecorder();
+      final ui.Canvas canvas = ui.Canvas(testRecorder);
+      canvas.drawRect(rect, ui.Paint()..color = const ui.Color.fromARGB(255, 255, 0, 0));
+      return testRecorder.endRecording();
+    }
+
+    final ui.Picture picture = rectPicture(const ui.Rect.fromLTWH(0, 0, 20, 20));
+
+    await createPlatformView(0, 'test-view');
+
+    final ui.SceneBuilder sb = ui.SceneBuilder();
+    sb.addPicture(ui.Offset.zero, picture);
+    sb.addPlatformView(0, width: 20, height: 20);
+    final ui.Scene scene = sb.build();
+    await renderScene(scene);
+
+    picture.dispose();
+    final Map<String, int> debugCounters = Instrumentation.instance.debugCounters;
+    expect(debugCounters['Picture Created'], debugCounters['Picture Deleted']);
+  });
 }
 
 // Returns the list of CSS transforms applied to the ancestor chain of
