+++ engine/src/flutter/lib/web_ui/test/ui/utils.dart
@@ -3,6 +3,7 @@
 // found in the LICENSE file.
 
 import 'dart:async';
+import 'dart:typed_data';
 
 import 'package:ui/src/engine.dart';
 import 'package:ui/src/engine/skwasm/skwasm_impl.dart'
@@ -24,6 +25,47 @@ Future<void> drawPictureUsingCurrentRenderer(Picture picture) async {
   sb.pushOffset(0, 0);
   sb.addPicture(Offset.zero, picture);
   await renderScene(sb.build());
+}
+
+// Sends a platform message to create a Platform View with the given id and viewType.
+Future<void> createPlatformView(int id, String viewType) {
+  final Completer<void> completer = Completer<void>();
+  const MethodCodec codec = StandardMethodCodec();
+  PlatformDispatcher.instance.sendPlatformMessage(
+    'flutter/platform_views',
+    codec.encodeMethodCall(MethodCall('create', <String, dynamic>{'id': id, 'viewType': viewType})),
+    (dynamic _) => completer.complete(),
+  );
+  return completer.future;
+}
+
+/// Disposes of the platform view with the given [id].
+Future<void> disposePlatformView(int id) {
+  final Completer<void> completer = Completer<void>();
+  const MethodCodec codec = StandardMethodCodec();
+  PlatformDispatcher.instance.sendPlatformMessage(
+    'flutter/platform_views',
+    codec.encodeMethodCall(MethodCall('dispose', id)),
+    (dynamic _) => completer.complete(),
+  );
+  return completer.future;
+}
+
+Future<bool> matchImage(Image left, Image right) async {
+  if (left.width != right.width || left.height != right.height) {
+    return false;
+  }
+  int getPixel(ByteData data, int x, int y) => data.getUint32((x + y * left.width) * 4);
+  final ByteData leftData = (await left.toByteData())!;
+  final ByteData rightData = (await right.toByteData())!;
+  for (int y = 0; y < left.height; y++) {
+    for (int x = 0; x < left.width; x++) {
+      if (getPixel(leftData, x, y) != getPixel(rightData, x, y)) {
+        return false;
+      }
+    }
+  }
+  return true;
 }
 
 /// Convenience getter for the implicit view.
